+++ src/main/java/com/saintdan/framework/vo/ClientVO.java
@@ -2,17 +2,18 @@ package com.saintdan.framework.vo;
 
 import org.springframework.security.core.GrantedAuthority;
 
+import java.io.Serializable;
 import java.util.Collection;
 import java.util.Set;
 
 /**
- * Client's VO.
+ * VO for {@link com.saintdan.framework.po.Client}
  *
  * @author <a href="http://github.com/saintdan">Liao Yifan</a>
  * @date 10/25/15
  * @since JDK1.8
  */
-public class ClientVO extends ResultVO {
+public class ClientVO implements Serializable {
 
     private static final long serialVersionUID = -6088897333280284093L;