+++ src/main/java/com/saintdan/framework/vo/LogVO.java
@@ -2,16 +2,17 @@ package com.saintdan.framework.vo;
 
 import com.saintdan.framework.enums.LogType;
 
+import java.io.Serializable;
 import java.util.Date;
 
 /**
- * Log's VO.
+ * VO for {@link com.saintdan.framework.po.Log}
  *
  * @author <a href="http://github.com/saintdan">Liao Yifan</a>
  * @date 10/27/15
  * @since JDK1.8
  */
-public class LogVO extends ResultVO {
+public class LogVO implements Serializable {
 
     private static final long serialVersionUID = -8802363013216964724L;