+++ src/main/java/com/saintdan/framework/vo/PageVO.java
@@ -2,14 +2,16 @@ package com.saintdan.framework.vo;
 
 import org.springframework.data.domain.Page;
 
+import java.io.Serializable;
+
 /**
- * Page's VO.
+ * VO for {@link Page}
  *
  * @author <a href="http://github.com/saintdan">Liao Yifan</a>
  * @date 10/27/15
  * @since JDK1.8
  */
-public class PageVO extends ResultVO {
+public class PageVO implements Serializable {
 
     private static final long serialVersionUID = 2808529039303273022L;