+++ src/main/java/com/saintdan/framework/vo/ResourceVO.java
@@ -1,13 +1,15 @@
 package com.saintdan.framework.vo;
 
+import java.io.Serializable;
+
 /**
- * Resource's VO.
+ * VO for {@link com.saintdan.framework.po.Resource}
  *
  * @author <a href="http://github.com/saintdan">Liao Yifan</a>
  * @date 10/16/15
  * @since JDK1.8
  */
-public class ResourceVO extends ResultVO {
+public class ResourceVO implements Serializable {
 
     private static final long serialVersionUID = 8917291426920312742L;