+++ src/main/java/com/saintdan/framework/vo/ResultVO.java
@@ -21,6 +21,8 @@ public class ResultVO implements Serializable {
 
     private String message;
 
+    private Object data;
+
     public ResultVO() {
 
     }
@@ -57,5 +59,13 @@ public class ResultVO implements Serializable {
 
     public void setMessage(String message) {
         this.message = message;
+    }
+
+    public Object getData() {
+        return data;
+    }
+
+    public void setData(Object data) {
+        this.data = data;
     }
 }